Update imports. Separate appender/logger so JSON output is without logger prefixes. Add Console appender
