+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/mvc/controller/MarkupController.java
@@ -223,12 +223,6 @@ public class MarkupController {
             response.flushBuffer();
             return;
         }
-
-        // Add audit logging
-        auditEventLogger.log(LogAuditEventRecord.TagType.SECURITY, 
-                            LogAuditEventRecord.OpType.READ, 
-                            LogAuditEventRecord.ResType.ACCESS, 
-                            "Downloaded markup file: id=" + id + ", uri=" + markupResult.getMarkupUri());
         
         Path localPath = IoUtils.toLocalPath(markupResult.getMarkupUri()).orElse(null);
         if (localPath != null) {
@@ -239,6 +233,11 @@ public class MarkupController {
                 return;
             }
             ioUtils.sendBinaryResponse(localPath, response);
+            
+            auditEventLogger.log(LogAuditEventRecord.TagType.SECURITY, 
+                                LogAuditEventRecord.OpType.EXTRACT, 
+                                LogAuditEventRecord.ResType.ALLOW, 
+                                "Downloaded markup file: uri=" + markupResult.getMarkupUri());
             return;
         }
 
@@ -276,6 +275,11 @@ public class MarkupController {
                     var s3Response = s3Stream.response();
                     IoUtils.sendBinaryResponse(s3Stream, response,
                             s3Response.contentType(), s3Response.contentLength());
+                    // Log after successful S3 download
+                    auditEventLogger.log(LogAuditEventRecord.TagType.SECURITY, 
+                                        LogAuditEventRecord.OpType.EXTRACT, 
+                                        LogAuditEventRecord.ResType.ALLOW, 
+                                        "Downloaded markup file: uri=" + markupResult.getMarkupUri());
                 }
                 return;
             } catch (StorageException e) {
@@ -292,6 +296,11 @@ public class MarkupController {
             try (InputStream inputStream = urlConnection.getInputStream()) {
                 IoUtils.sendBinaryResponse(inputStream, response, urlConnection.getContentType(),
                         urlConnection.getContentLength());
+                // Log after successful URL download
+                auditEventLogger.log(LogAuditEventRecord.TagType.SECURITY, 
+                                    LogAuditEventRecord.OpType.EXTRACT, 
+                                    LogAuditEventRecord.ResType.ALLOW, 
+                                    "Downloaded markup file: uri=" + markupResult.getMarkupUri());
             }
         } catch (IOException e) {
             log.error("Markup with id " + id + " download failed: " + e.getMessage());
