+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/mvc/security/LocalSecurityConfig.java
@@ -53,7 +53,7 @@ public class LocalSecurityConfig {
     @Bean
     public AuthenticationSuccessHandler authenticationSuccessHandler() {
         return (request, response, authentication) -> {
-            auditEventLogger.log(LogAuditEventRecord.TagType.SECURITY, LogAuditEventRecord.OpType.LOGIN, LogAuditEventRecord.ResType.ACCESS, "User successfully logged in.");
+            auditEventLogger.log(LogAuditEventRecord.TagType.SECURITY, LogAuditEventRecord.OpType.LOGIN, LogAuditEventRecord.ResType.ALLOW, "User successfully logged in.");
             response.sendRedirect("/");
         };
     }
